pyproject.toml
[project]
name = "rag-faq-plurelo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "haystack-ai>=2.17.1",
    "isort>=6.0.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=4.1.0",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
]

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 89
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
